Resolved weird issue with SPDP_10.c
There was a weird issue with upper-case/lower-case variants of
spdp_10.c, whereas the Makefile used the lower-case variant only.
Whenever the repository was cloned, git status showed that SPDP_10.c was
modified without having touched the file (and it did not even exist,
only the lower-case spdp_10.c variant exists).  I have merged the file
content of SPDP_10.c into spdp_10.c (as it is used in the Makefile) and
deleted SPDP_10.c.